# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Portfolio Manager API
  summary: 股票与 ETF 投资组合管理 REST API
  description: |
    管理投资组合、股票/ETF 交易、每日行情同步、组合估值和资产配置。
    MVP 使用服务端预置演示用户，金融数值以十进制字符串传输。
  version: 1.0.0
  contact:
    name: Portfolio Manager Team
servers:
  - url: /
    description: 当前部署地址；本地开发默认解析为 http://localhost:8000
security: []
...
paths:
  /api/v1/portfolios:
    get:
      tags: [Portfolios]
      operationId: listPortfolios
      summary: 查询组合列表
      description: 查询当前用户的组合，默认不包含已归档组合。
      x-owner: member-2
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: includeArchived
          in: query
          required: false
          schema:
            type: boolean
            default: false
          example: false
      responses:
        "200":
          description: 组合分页列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioPage"
              example:
                items:
                  - id: 22222222-2222-2222-2222-222222222222
                    name: Long-term Growth
                    description: US stocks and ETFs
                    baseCurrency: USD
                    isArchived: false
                    createdAt: "2026-07-27T08:00:00Z"
                    updatedAt: "2026-07-27T08:00:00Z"
                page: 1
                pageSize: 20
                total: 1
        "422":
          $ref: "#/components/responses/ValidationError"
    post:
      tags: [Portfolios]
      operationId: createPortfolio
      summary: 创建组合
      x-owner: member-2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortfolioCreate"
            example:
              name: Long-term Growth
              description: US stocks and ETFs
              baseCurrency: USD
      responses:
        "201":
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Portfolio"
              example:
                id: 22222222-2222-2222-2222-222222222222
                name: Long-term Growth
                description: US stocks and ETFs
                baseCurrency: USD
                isArchived: false
                createdAt: "2026-07-27T08:00:00Z"
                updatedAt: "2026-07-27T08:00:00Z"
        "409":
          $ref: "#/components/responses/ConflictError"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/portfolios/{portfolioId}:
    parameters:
      - $ref: "#/components/parameters/PortfolioId"
    get:
      tags: [Portfolios]
      operationId: getPortfolio
      summary: 查询组合详情
      x-owner: member-2
      responses:
        "200":
          description: 组合详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Portfolio"
              example:
                id: 22222222-2222-2222-2222-222222222222
                name: Long-term Growth
                description: US stocks and ETFs
                baseCurrency: USD
                isArchived: false
                createdAt: "2026-07-27T08:00:00Z"
                updatedAt: "2026-07-27T08:00:00Z"
        "404":
          $ref: "#/components/responses/NotFoundError"
    patch:
      tags: [Portfolios]
      operationId: updatePortfolio
      summary: 修改组合
      x-owner: member-2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortfolioUpdate"
            example:
              name: Core Growth
              description: Updated description
      responses:
        "200":
          description: 修改后的组合
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Portfolio"
              example:
                id: 22222222-2222-2222-2222-222222222222
                name: Core Growth
                description: Updated description
                baseCurrency: USD
                isArchived: false
                createdAt: "2026-07-27T08:00:00Z"
                updatedAt: "2026-07-27T09:00:00Z"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "422":
          $ref: "#/components/responses/ValidationError"
    delete:
      tags: [Portfolios]
      operationId: deletePortfolio
      summary: 删除空组合
      description: 仅允许删除没有交易历史的组合。
      x-owner: member-2
      responses:
        "204":
          description: 删除成功，无响应体
        "404":
          $ref: "#/components/responses/NotFoundError"
        "409":
          $ref: "#/components/responses/ConflictError"
...
components:
  parameters:
    PortfolioId:
      name: portfolioId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      example: 22222222-2222-2222-2222-222222222222
...
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      example: 1
    PageSize:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      example: 20
...
  responses:
    NotFoundError:
      description: 资源不存在或不属于当前用户
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: PORTFOLIO_NOT_FOUND
            message: Portfolio was not found.
            fieldErrors: {}
            requestId: req-01J3X6F8NQ8H2A3V
    ConflictError:
      description: 唯一键、幂等、并发或业务状态冲突
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: POSITION_INSUFFICIENT_QUANTITY
            message: 卖出数量超过当前持仓。
            fieldErrors:
              quantity:
                - Must not exceed the current position.
            requestId: req-01J3X6F8NQ8H2A3V
    ValidationError:
      description: 请求字段验证失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: VALIDATION_ERROR
            message: Request validation failed.
            fieldErrors:
              quantity:
                - Must be greater than zero.
            requestId: req-01J3X6F8NQ8H2A3V
...
  schemas:
...
    Currency:
      type: string
      pattern: "^[A-Z]{3}$"
      example: USD
...
    ErrorResponse:
      type: object
      required: [code, message, fieldErrors, requestId]
      properties:
        code:
          type: string
        message:
          type: string
        fieldErrors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
        requestId:
          type: string
    PortfolioCreate:
      type: object
      required: [name, baseCurrency]
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: [string, "null"]
          maxLength: 500
        baseCurrency:
          $ref: "#/components/schemas/Currency"
      example:
        name: Long-term Growth
        description: US stocks and ETFs
        baseCurrency: USD
    PortfolioUpdate:
      type: object
      minProperties: 1
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: [string, "null"]
          maxLength: 500
      example:
        name: Core Growth
        description: Updated description
    Portfolio:
      type: object
      required:
        - id
        - name
        - baseCurrency
        - isArchived
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: [string, "null"]
        baseCurrency:
          $ref: "#/components/schemas/Currency"
        isArchived:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PortfolioPage:
      type: object
      required: [items, page, pageSize, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Portfolio"
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
...